chore: rename items::component to primitive
