+++ src/items/populate_items.rs
@@ -7,14 +7,14 @@ use crate::items::bundle::lever_bundle::LeverBundle;
 use crate::items::bundle::pressure_plate_bundle::PressurePlateBundle;
 use crate::items::bundle::pressure_plate_on_off_bundle::PressurePlateOnOffBundle;
 use crate::items::bundle::teleporter_bundle::TeleporterBundle;
-use crate::items::components::dependencies::{Dependencies, Off, On};
-use crate::items::components::ghost_only::GhostOnly;
-use crate::items::components::killing::Killing;
-use crate::items::components::level_teleporter::LevelTeleporter;
-use crate::items::components::player_only::PlayerOnly;
-use crate::items::components::single_use::SingleUse;
-use crate::items::components::start_timer::StartTimer;
-use crate::items::components::teleporter::Teleporter;
+use crate::items::primitive::dependencies::{Dependencies, Off, On};
+use crate::items::primitive::ghost_only::GhostOnly;
+use crate::items::primitive::killing::Killing;
+use crate::items::primitive::level_teleporter::LevelTeleporter;
+use crate::items::primitive::player_only::PlayerOnly;
+use crate::items::primitive::single_use::SingleUse;
+use crate::items::primitive::start_timer::StartTimer;
+use crate::items::primitive::teleporter::Teleporter;
 use crate::map_parser::map_repr::{InteractionType, ObjectRepr, ObjectType};
 use crate::math::vec2i::Vec2i;
 use bevy::prelude::*;
