+++ src/items/primitive/is_activated.rs
