+++ src/items/systems/timer_system.rs
@@ -1,7 +1,7 @@
-use crate::items::components::cooldown::Cooldown;
-use crate::items::components::is_activated::IsActivated;
-use crate::items::components::is_usable::IsUsable;
-use crate::items::components::start_timer::StartTimer;
+use crate::items::primitive::cooldown::Cooldown;
+use crate::items::primitive::is_activated::IsActivated;
+use crate::items::primitive::is_usable::IsUsable;
+use crate::items::primitive::start_timer::StartTimer;
 use bevy::prelude::*;
 
 pub fn cooldown_system(
